initiate utils._context._scenarios.module
